Converting Quartz audio frame counting to be double instead of integer - the latter causes clock drift over time and causes issue when synchronizing with external sources or long audio
#jira UE-147371
#rb Aaron.McLeran
#preflight 624600f5e434babd8af9bcc4

#ROBOMERGE-AUTHOR: maxwell.hayes
#ROBOMERGE-SOURCE: CL 19578173 via CL 19579759 via CL 19579771 via CL 19579784
#ROBOMERGE-BOT: UE5 (Release-Engine-Staging -> Main) (v938-19570697)

[CL 19580686 by maxwell hayes in ue5-main branch]

+++ Engine/Source/Runtime/AudioMixer/Private/Quartz/AudioMixerQuantizedCommands.cpp
@@ -119,7 +119,7 @@ namespace Audio
 	int32 FQuantizedQueueCommand::OverrideFramesUntilExec(int32 NumFramesUntilExec)
 	{
 		// Calculate the amount of time before taking up a voice slot
-		int32 NumFramesBeforeVoiceSlot = NumFramesUntilExec - OwningClockPtr->GetTickRate().GetFramesPerDuration(AudioComponentData.AnticapatoryBoundary.Quantization);
+		int32 NumFramesBeforeVoiceSlot = NumFramesUntilExec - static_cast<int32>(OwningClockPtr->GetTickRate().GetFramesPerDuration(AudioComponentData.AnticapatoryBoundary.Quantization));
 
 		//If NumFramesBeforeVoiceSlot is less than 0, change the boundary back to the original, and mark this command as having 0 frames till exec
 		if (NumFramesBeforeVoiceSlot < 0)
